actix_ws

Struct Session

pub struct Session { /* private fields */ }

A handle into the websocket session.

This type can be used to send messages into the WebSocket.

Implementations

impl Session

pub async fn text(&mut self, msg: impl Into<ByteString>) -> Result<(), Closed>

Sends text into the WebSocket.

if session.text("Some text").await.is_err() {
    // session closed
}

pub async fn binary(&mut self, msg: impl Into<Bytes>) -> Result<(), Closed>

Sends raw bytes into the WebSocket.

if session.binary(&b"some bytes"[..]).await.is_err() {
    // session closed
}

pub async fn ping(&mut self, msg: &[u8]) -> Result<(), Closed>

Pings the client.

For many applications, it will be important to send regular pings to keep track of if the client has disconnected

if session.ping(b"").await.is_err() {
    // session is closed
}

pub async fn pong(&mut self, msg: &[u8]) -> Result<(), Closed>

Pongs the client.

match msg {
    Message::Ping(bytes) => {
        let _ = session.pong(&bytes).await;
    }
    _ => (),
}

pub async fn continuation(&mut self, msg: Item) -> Result<(), Closed>

Manually controls sending continuations.

Be wary of this method. Continuations represent multiple frames that, when combined, are presented as a single message. They are useful when the entire contents of a message are not available all at once. However, continuations MUST NOT be interrupted by other Text or Binary messages. Control messages such as Ping, Pong, or Close are allowed to interrupt a continuation.

Continuations must be initialized with a First variant, and must be terminated by a Last variant, with only Continue variants sent in between.

session.continuation(Item::FirstText("Hello".into())).await?;
session.continuation(Item::Continue(b", World"[..].into())).await?;
session.continuation(Item::Last(b"!"[..].into())).await?;

pub async fn close(self, reason: Option<CloseReason>) -> Result<(), Closed>

Sends a close message, and consumes the session.

All clones will return Err(Closed) if used after this call.

session.close(None).await